refactor: everything is correctly queried and tracked
