+++ crates/biome_css_analyze/src/lint/nursery/no_unused_classes.rs
@@ -2,7 +2,7 @@ use crate::services::module_graph::CssModuleGraph;
 use biome_analyze::{Rule, RuleDiagnostic, RuleDomain, context::RuleContext, declare_lint_rule};
 use biome_console::markup;
 use biome_css_syntax::{CssClassSelector, CssPseudoClassFunctionSelector};
-use biome_module_graph::is_class_referenced_by_importers;
+use biome_module_graph::{ModuleDb, SymbolName, is_class_referenced_by_importers};
 use biome_rowan::AstNode;
 use biome_rule_options::no_unused_classes::NoUnusedClassesOptions;
 
@@ -74,12 +74,15 @@ impl Rule for NoUnusedClasses {
         let class_name = node.name().ok()?;
         let class_name = class_name.as_css_custom_identifier()?;
         let class_name = class_name.value_token().ok()?;
-        let class_name_text = class_name.text_trimmed();
 
         let db = ctx.db();
-        let file_path = ctx.file_path();
+        let module = db.module_for_path(ctx.file_path())?;
 
-        if is_class_referenced_by_importers(db, file_path, class_name_text) {
+        if is_class_referenced_by_importers(
+            db,
+            module,
+            SymbolName::new(db, class_name.token_text_trimmed().text()),
+        ) {
             return None;
         }
