+++ crates/biome_html_analyze/src/lint/nursery/no_undeclared_classes.rs
@@ -142,8 +142,8 @@ impl Rule for NoUndeclaredClasses {
             }
 
             // Only build the import tree for diagnostic display when the class is missing.
-            if !found_class {
-                let import_tree = build_import_tree_for_html(db, file_path);
+            if !found_class && let Some(module) = db.module_for_path(file_path) {
+                let import_tree = build_import_tree_for_html(db, module);
                 let start = TextSize::from(class_data.inner_file_start + class_offset);
                 let end = start + TextSize::from(class_name.len() as u32);
                 signals.push(UndeclaredClass {
