+++ crates/biome_js_analyze/src/lint/suspicious/no_deprecated_imports.rs
@@ -4,8 +4,10 @@ use biome_analyze::{
 use biome_console::markup;
 use biome_diagnostics::Severity;
 use biome_js_syntax::{AnyJsImportClause, AnyJsImportLike, JsModuleSource};
-use biome_module_graph::{JsImportPath, JsModuleInfo, ModuleDb};
-use biome_rowan::{AstNode, TextRange};
+use biome_module_graph::{
+    JsImportPath, ModuleDb, ModuleInfo, SymbolName, find_jsdoc_for_exported_symbol,
+};
+use biome_rowan::{AstNode, Text, TextRange};
 use biome_rule_options::no_deprecated_imports::NoDeprecatedImportsOptions;
 use camino::Utf8Path;
 
@@ -125,7 +127,7 @@ fn get_deprecated_imports_from_module_source(
     target_path: &Utf8Path,
     module_db: &dyn ModuleDb,
 ) -> Vec<NoDeprecatedImportsState> {
-    let Some(module_info) = module_db.js_module_info_for_path(target_path) else {
+    let Some(module_info) = module_db.module_for_path(target_path) else {
         return Vec::new();
     };
 
@@ -134,7 +136,7 @@ fn get_deprecated_imports_from_module_source(
     };
 
     import_clause.filter_map_all_imported_symbols(|name, range| {
-        find_deprecation(&module_info, module_db, &name)
+        find_deprecation(module_info, module_db, &name)
             .map(|message| NoDeprecatedImportsState { range, message })
     })
 }
@@ -147,37 +149,40 @@ fn get_deprecated_imports_from_module_source(
 /// - Returns `Some(None)` if the symbol is deprecated but has no message.
 /// - Returns `None` if the symbol is not deprecated or cannot be found.
 fn find_deprecation(
-    module_info: &JsModuleInfo,
+    module_info: ModuleInfo,
     module_db: &dyn ModuleDb,
-    name: &str,
+    name: &Text,
 ) -> Option<Option<String>> {
-    module_info
-        .find_jsdoc_for_exported_symbol(module_db, name)
-        .and_then(|jsdoc| {
-            let mut is_deprecated = false;
-            let mut message = String::new();
-            for line in jsdoc.lines() {
-                let line = line.trim();
-                if is_deprecated {
-                    if line.is_empty() {
-                        break;
-                    }
-
-                    if !message.is_empty() {
-                        message.push(' ');
-                    }
-
-                    message.push_str(line);
-                } else if let Some((_before, after)) = line.split_once("@deprecated") {
-                    is_deprecated = true;
-                    message.push_str(after.trim_start());
+    find_jsdoc_for_exported_symbol(
+        module_db,
+        module_info,
+        SymbolName::new(module_db, name.text()),
+    )
+    .and_then(|jsdoc| {
+        let mut is_deprecated = false;
+        let mut message = String::new();
+        for line in jsdoc.lines() {
+            let line = line.trim();
+            if is_deprecated {
+                if line.is_empty() {
+                    break;
                 }
-            }
 
-            is_deprecated.then_some(if message.is_empty() {
-                None
-            } else {
-                Some(message)
-            })
+                if !message.is_empty() {
+                    message.push(' ');
+                }
+
+                message.push_str(line);
+            } else if let Some((_before, after)) = line.split_once("@deprecated") {
+                is_deprecated = true;
+                message.push_str(after.trim_start());
+            }
+        }
+
+        is_deprecated.then_some(if message.is_empty() {
+            None
+        } else {
+            Some(message)
         })
+    })
 }
